+++ order-processor/pom.xml
@@ -8,7 +8,7 @@
 		<version>2.5.6</version>
 		<relativePath/> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>io.github.smallintro</groupId>
+	<groupId>io.github.smallintro.rabbitmq</groupId>
 	<artifactId>order-processor</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>order-processor</name>
