+++ order-processor/src/main/java/io/github/smallintro/rabbitmq/orderprocessor/OrderProcessorApplication.java
@@ -1,4 +1,4 @@
-package io.github.smallintro.orderprocessor;
+package io.github.smallintro.rabbitmq.orderprocessor;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
