+++ order-processor/src/main/java/io/github/smallintro/rabbitmq/orderprocessor/model/PaymentStatus.java
@@ -1,6 +1,7 @@
-package io.github.smallintro.orderprocessor.model;
+package io.github.smallintro.rabbitmq.orderprocessor.model;
 
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -13,8 +14,15 @@ import lombok.ToString;
 @EqualsAndHashCode
 @ToString
 public class PaymentStatus {
+    @JsonProperty("payment_request_id")
     private String paymentRequestId;
+
+    @JsonProperty("payment_status")
     private String paymentStatus;
+
+    @JsonProperty("message")
     private String message;
+
+    @JsonProperty("payment_info")
     private PaymentInfo paymentInfo;
 }
