--- order-processor/src/main/java/io/github/smallintro/rabbitmq/paymentprocessor/PaymentProcessorApplication.java
@@ -1,13 +1,0 @@
-package io.github.smallintro.rabbitmq.paymentprocessor;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PaymentProcessorApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PaymentProcessorApplication.class, args);
-	}
-
-}
