--- order-processor/src/test/java/io/github/smallintro/rabbitmq/paymentprocessor/PaymentProcessorApplicationTests.java
@@ -1,13 +1,0 @@
-package io.github.smallintro.rabbitmq.paymentprocessor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentProcessorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
